# cmd/producer/docs/openapi.yaml
openapi: 3.0.0
info:
  title: Task Processor API
  description: API for creating tasks and sending them to RabbitMQ.
  version: "1.0"
paths:
  /tasks:
    get:
      summary: Get all tasks
      description: Get tasks for all statuses.
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"
        "400":
          description: Failed to retrieve tasks from MongoDB
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
    post:
      summary: Create a new task
      description: Create a new task and send it to the RabbitMQ queue.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskRequest"
      responses:
        "201":
          description: Task created successfully
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
servers:
  - url: //localhost:8080
components:
  schemas:
    Task:
      description: Represents a task with dynamic payloads based on the task type.
      type: object
      properties:
        id:
          type: integer
        type:
          type: string
          example: email
        payload:
          oneOf:
            - $ref: "#/components/schemas/EmailPayload"
            - $ref: "#/components/schemas/HealthCheckPayload"
        createdAt:
          type: string
          format: date-time
        status:
          type: string
        finishedAt:
          type: string
          format: date-time
        result:
          $ref: "#/components/schemas/TaskResult"
    TaskRequest:
      description: Represents a task request. The payload type depends on the task type.
      type: object
      properties:
        type:
          type: string
          example: email
        payload:
          oneOf:
            - $ref: "#/components/schemas/EmailPayload"
            - $ref: "#/components/schemas/HealthCheckPayload"
    TaskResult:
      type: object
      properties:
        status:
          type: string
        error:
          type: string
        detail:
          type: string
        timestamp:
          type: string
          format: date-time
    EmailPayload:
      type: object
      properties:
        email:
          type: string
        subject:
          type: string
        body:
          type: string
    HealthCheckPayload:
      type: object
      properties:
        url:
          type: string
        method:
          type: string
        status:
          type: string